refactor [윤아영] : 타입 파일 변경, import시 타입 파일임을 명시
