+++ apis/request.ts
@@ -1,7 +1,7 @@
+import type { Members } from '@/components/tables/members.type';
 import axios from './axios';
 import getToken from './localStorage';
 import { DashboardColors } from '@/components/dashboard/dashboard.constants';
-import { Members } from '@/components/tables/members-type';
 
 const ERROR_MESSAGE = '에러 발생:';
